test: move tests around & fix
